+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Release 2.15.2](https://github.com/aws/aws-sdk-android/releases/tag/release_v2.15.2)
 
+### Bug Fixes
+- **AWS Mobile Client**
+  - Fix issue where MobileClient would return invalid tokens after initialization when using Hosted UI.
+    See [issue #873](https://github.com/aws-amplify/aws-sdk-android/issues/873) and
+    [issue #888](https://github.com/aws-amplify/aws-sdk-android/issues/888)
+
 ### Misc. Updates
 
 - Model updates for the following services